+++ src/A21_update.cpp
@@ -13,26 +13,30 @@ Rcpp::List A21_update(double A21_old,
                       arma::vec delta2,
                       double sigma2_A,
                       double metrop_var_A21,
-                      double acctot_A21){
+                      int acctot_A21){
 
 /*Second*/
-arma::vec mean_piece_old = gamma_star - A21_old*delta1 - A22*delta2;
+arma::vec mean_piece_old = gamma_star - 
+                           A21_old*delta1 - 
+                           A22*delta2;
 
 double second = -0.50*dot(mean_piece_old, mean_piece_old) - 
-                0.50*(1/sigma2_A)*(A21_old*A21_old);
+                0.50*(1.00/sigma2_A)*(A21_old*A21_old);
 
 /*First*/
 double A21 = R::rnorm(A21_old, 
                       sqrt(metrop_var_A21));
-arma::vec mean_piece = gamma_star - A21*delta1 - A22*delta2;
+arma::vec mean_piece = gamma_star - 
+                       A21*delta1 - 
+                       A22*delta2;
 
 double first = -0.50*dot(mean_piece, mean_piece) - 
-               0.50*(1/sigma2_A)*(A21*A21);
+               0.50*(1.00/sigma2_A)*(A21*A21);
 
 /*Decision*/
 double ratio = exp(first - second);   
-double acc = 1;
-if(ratio < R::runif(0, 1)){
+int acc = 1;
+if(ratio < R::runif(0.00, 1.00)){
   A21 = A21_old;
   acc = 0;
   }
